Moved functions to related files
